+++ migrations/20260801011525_create_refunds.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS refunds (
   id UUID PRIMARY KEY,
   donation_id UUID UNIQUE NOT NULL REFERENCES donations(id),
   event_id UUID UNIQUE NOT NULL REFERENCES blockchain_events(id),
-  amount INTEGER NOT NULL CHECK (amount > 0),
+  amount BIGINT NOT NULL CHECK (amount > 0),
   created_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
